refactor(firecracker): move semgrep guest + guest-init to projects/firecracker/semgrep
