+++ bazel/images/BUILD
@@ -46,9 +46,9 @@ multirun(
         "//projects/agent_platform/fc-agentd/rootfs-builder:image.push",
         "//projects/agent_platform/fc-agentd:image.push",
         "//projects/agent_platform/harness:image.push",
-        "//projects/agent_platform/semgrep-guest:image.push",
         "//projects/agent_platform/semgrep-scand/chart:chart.push",
         "//projects/agent_platform/semgrep-scand:image.push",
+        "//projects/firecracker/semgrep/guest:image.push",
         "//projects/firecracker/substrate/egress-proxy:image.push",
         "//projects/grimoire/api:image.push",
         "//projects/grimoire/frontend:image.push",
