+++ projects/firecracker/semgrep/guest-init/cmd/loopback_other.go
