+++ projects/firecracker/semgrep/guest-init/cmd/main.go
@@ -22,8 +22,8 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/jomcgi/homelab/projects/agent_platform/semgrep-guest-init/internal/lspdriver"
-	"github.com/jomcgi/homelab/projects/agent_platform/semgrep-guest-init/internal/scanserver"
+	"github.com/jomcgi/homelab/projects/firecracker/semgrep/guest-init/internal/lspdriver"
+	"github.com/jomcgi/homelab/projects/firecracker/semgrep/guest-init/internal/scanserver"
 	"github.com/jomcgi/homelab/projects/firecracker/substrate/vsockproto"
 )
