+++ projects/firecracker/semgrep/guest-init/cmd/mount_other.go
