+++ projects/firecracker/semgrep/guest-init/internal/lspdriver/BUILD
@@ -3,8 +3,8 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "lspdriver",
     srcs = ["lspdriver.go"],
-    importpath = "github.com/jomcgi/homelab/projects/agent_platform/semgrep-guest-init/internal/lspdriver",
-    visibility = ["//projects/agent_platform/semgrep-guest-init:__subpackages__"],
+    importpath = "github.com/jomcgi/homelab/projects/firecracker/semgrep/guest-init/internal/lspdriver",
+    visibility = ["//projects/firecracker/semgrep/guest-init:__subpackages__"],
     deps = ["//projects/firecracker/substrate/vsockproto"],
 )
